Fungal garden data

Fungal gardens in fungus-growing ants function as an external digestive system, where mutualistic fungi of the genus Leucoagaricus convert plant substrate into energy. In this study, a naturally evolved fungal garden consortium underwent gradual infection by a pathogenic fungus, Escovopsis, creating a dynamic system for investigation. The researchers analyzed the bacterial taxa in this system to track community shifts and spatial activity mapping.

Format

two datasets: fungal_garden and pathway

Source

Generated from PNNL

Details

As microorganisms can provide defense by competitively excluding pathogenic microbes, a central question was to uncover changes in population compositions and metabolic potentials within this intricate microbial consortium in response to perturbation. To grasp this matter, we can employ the MolPad dashboard to examine a particular pathway and unveil its pattern across various subtypes.

Fungal garden data

A list that contains 3 dataframes:

ID

sample ID

D0_01~D10_11

measured values in the corresponding days

Pathway

A data frame with 126725 rows and 4 variables:

ID

sample ID

taxonomic.scope

the category of the feature

ko_term

KEGG IDs, linking genomic information with higher order functional information.

Pathway

KEGG pathways, representation focuses on the network of gene products, mostly proteins but including functional RNAs.

Examples

data(fungal_garden)
head(fungal_dlist[[1]],3)
#>       ID      D0_01     D1_02      D2_03     D3_04     D4_05     D5_06
#> 1 grp1_1 -0.9533405 -1.106882 -0.7821818 -2.344134 -2.213369 -1.846307
#> 2 grp1_2 -0.9533405 -1.106882 -0.7821818 -1.740328 -1.983804 -1.910560
#> 3 grp1_6  0.9672390  1.351778  1.2871986  1.306616  1.253802  1.228330
#>       D6_07     D7_08     D8_09     D9_10    D10_11
#> 1 -2.019264 -1.653271 -1.855675 -1.856335 -2.675131
#> 2 -2.360466 -1.728546 -2.204190 -2.552709 -3.220536
#> 3  1.522306  1.381057  1.512031  1.391325  1.297332
head(fungal_dlist[[2]],3)
#>       ID       D0_01       D1_02        D2_03       D3_04        D4_05
#> 1 grp2_1  0.07955562 -0.06847126 -0.085804628 -0.06970496 -0.189086831
#> 2 grp2_2  0.14481731  0.63121925  0.640259446  0.53728726  0.003337022
#> 3 grp2_3 -0.71104811 -0.03273029  0.007307968  0.15272255  0.151116165
#>         D5_06       D6_07      D7_08      D8_09      D9_10     D10_11
#> 1 -0.01056197 -0.30973183 -0.1719941 -0.1238594 -0.2293862  0.1021993
#> 2 -0.20361189  0.62014296  0.9765937  1.0188160  0.8699408  0.3023357
#> 3  0.04532175  0.09048819  0.3211153  0.3079252  0.2960252 -0.4230718
head(fungal_dlist[[3]],3)
#>       ID      D0_01      D1_02     D2_03      D3_04      D4_05      D5_06
#> 1 grp3_1 -0.4702515 -0.3390523 -0.501346 -0.4833872 -0.4902185 -0.6680519
#> 2 grp3_2 -1.6644934 -2.6212720 -2.283416 -1.8775744 -0.6482506 -1.5863859
#> 3 grp3_3 -0.9248744 -0.6240496 -0.552006 -0.2821400 -0.3161406 -0.3980750
#>        D6_07      D7_08      D8_09      D9_10     D10_11
#> 1 -0.6609125 -0.7754728 -0.7022775 -0.5814545 -0.1418371
#> 2 -1.7470608 -0.3588664 -1.0980109 -1.7763473 -0.4198883
#> 3 -0.3106139 -0.2878704 -0.2226940 -0.1963857  0.4274527
head(pathway,3)
#>        ID taxonomic.scope ko_term              Pathway
#> 1 grp1_24        Bacteria  K02706 00195_Photosynthesis
#> 2 grp1_25        Bacteria  K02706 00195_Photosynthesis
#> 3 grp1_26        Bacteria  K02706 00195_Photosynthesis