Generate processed annotation

gAnnotation() provides the standard input format for the dashboard, allowing users to select two columns as the primary factors they wish to visualize and describe in the data.

Usage

gAnnotation(data, first_order, second_order)

Arguments

data

A dataframe.

first_order

The name of one column of categorical variable in data.

second_order

The name of another column of categorical variable in data.

Details

Guidelines for Selecting Annotations:

The first-order annotation is recommended to be functional, such as a pathway or functional system. This parameter will primarily serve the purpose of filtering out one network at a time when displaying the dashboard.

The second-order annotation is utilized to illustrate the composition of a first-order annotation using a bar plot. Therefore, it is better to set it as taxon, class label, etc.

Examples

data(test_data)
test_annotations_processed <- gAnnotation(test_annotations,system,class)
head(test_annotations_processed)
#>   ID
#> 1  1
#> 2  2
#> 3  3
#> 4  4
#> 5  5
#> 6  6
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     GO_ID
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
#> 3                                                                                                                                                                                                                                        GO:0003674,GO:0003824,GO:0005575,GO:0006629,GO:0006644,GO:0006650,GO:0006655,GO:0006793,GO:0006796,GO:0008150,GO:0008152,GO:0008610,GO:0008654,GO:0008808,GO:0009058,GO:0009987,GO:0016020,GO:0016740,GO:0016772,GO:0016780,GO:0019637,GO:0030572,GO:0032048,GO:0032049,GO:0044237,GO:0044238,GO:0044249,GO:0044255,GO:0045017,GO:0046471,GO:0046474,GO:0046486,GO:0071704,GO:0090407,GO:1901576
#> 4 GO:0003674,GO:0003676,GO:0003723,GO:0003729,GO:0003824,GO:0004654,GO:0005488,GO:0005575,GO:0005622,GO:0005623,GO:0005737,GO:0006139,GO:0006401,GO:0006402,GO:0006725,GO:0006807,GO:0008150,GO:0008152,GO:0009056,GO:0009057,GO:0009892,GO:0009987,GO:0010468,GO:0010605,GO:0010629,GO:0016070,GO:0016071,GO:0016740,GO:0016772,GO:0016779,GO:0019222,GO:0019439,GO:0034641,GO:0034655,GO:0043170,GO:0044237,GO:0044238,GO:0044248,GO:0044260,GO:0044265,GO:0044270,GO:0044424,GO:0044464,GO:0046483,GO:0046700,GO:0048519,GO:0050789,GO:0060255,GO:0065007,GO:0071704,GO:0090304,GO:0097159,GO:1901360,GO:1901361,GO:1901363,GO:1901575
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
#>   KEGG_ID               Pathway      taxonomic.scope
#> 1  K07124  Integumentary System    hormonal proteins
#> 2               Skeletal System  structural proteins
#> 3  K06131       Muscular System              enzymes
#> 4  K00962        Nervous System contractile proteins
#> 5  K02083      Endocrine System contractile proteins
#> 6         Cardiovascular System  structural proteins