Generate shiny dashboard

Once you have the outputs from the g-functions, you're ready to create your custom Molpad dashboard. Be sure to specify the web ID columns and their corresponding column names.

Usage

gDashboard(
  data,
  cluster,
  annotation,
  networkres,
  dashboardtitle = "MolPad Dashboard",
  id_colname = NULL,
  id_type = NULL
)

Arguments

data

The output of pre_process():a scaled data.frame contain variables ID, value on time_1, ..., value on time_k, and type.

cluster

The output of gClusters(): a list contains the result of k-means Cluster means, vectors and sum of squares.

annotation

The output of gAnnotation(): a data.frame containing annotations for describing the features. Variables must include ID, Pathway, and taxonomic.scope. Note that NA is not permitted for these 3 variables.

networkres

The output of gNetwork(): a data.frame with 4 variables weight,IncNodePurity,var_names,from.

dashboardtitle

A string of customized dashboard name.

id_colname

A single string or a string vector. The column names in your annotation dataset where the vector contains external database IDs.

id_type

A single string or a string vector. The corresponding database names for the above columns, must be chosen from "KEGG" and "GO".

Details

Please ensure that columns containing external database IDs adhere to the following standards:

KEGG ID: Begin with 'K' followed by 5 digits, for example, K05685 or K06671.

GO ID: Begin with 'GO:' followed by 7 digits, such as GO:0003674."

Examples

if (FALSE) { # \dontrun{
data(test_data)
gDashboard(test_data_processed,test_cluster,test_annotations_processed,test_network,id_colname = c("GO_ID","KEGG_ID"),id_type = c("GO","KEGG"))
} # }