Match color

Match any vector with a finite list of colors.

Usage

match.color__(valist, mycolors, extendby = 5)

Arguments

valist

A vector that you want to specify the colors for each element.

mycolors

A vector of colors.

extendby

A number to select from 1, 2, 3, 4, or 5, representing distinct auto-fill schemes.

Examples

my_vector <- paste0("N_",1:10)
match.color__(my_vector,c("red","yellow","blue"))
#>             N_1             N_2             N_3             N_4             N_5 
#>           "red"        "yellow"          "blue"         "white" "antiquewhite3" 
#>             N_6             N_7             N_8             N_9            N_10 
#>   "aquamarine3"        "azure3"       "bisque2"          "blue"    "blueviolet"