Cheese data

In the context of cheese production, the regular application of a brine solution is a maturation technique that promotes uniformity in the microbial populations on the cheese's surface and facilitates interactions among these microorganisms. This investigation involved the analysis of a longitudinal dataset encompassing three washed-rind cheese communities sampled during the ripening process.

Format

two datasets: cheese and annotations

Source

Reference: doi:10.1128/msystems.00701-22

Details

The study uncovered a remarkably consistent microbial progression within each cheese. In the bacterial community, Firmicutes dominate at the outset, with Proteobacteria swiftly assuming dominance by the end of the ripening period. Additionally, both Cheese A and Cheese C consistently demonstrate the establishment of Actinobacteria and Bacteroidetes, each in its own distinct manner. To corroborate these findings using the MolPad dashboard, we conducted an analysis of two cheeses (A and C) from all three production batches, spanning weeks 2 to 13.

cheese data

A data frame with 106239 rows and 18 variables:

ID

sample ID

A_1~A_5

Time series data measured for cheese A.

C_1~C_5

Time series data measured for cheese C.

domain

the category of the feature

phylum

the category of the feature

class

the category of the feature

order

the category of the feature

family

the category of the feature

genus

the category of the feature

annotation data

A data frame with 86156 rows and 9 variables:

ID

sample ID

GO_ID

GO IDs, represents a link between a gene product type and a molecular function

KEGG_ID

KEGG IDs, linking genomic information with higher order functional information.

domain

the category of the feature

phylum

the category of the feature

class

the category of the feature

order

the category of the feature

family

the category of the feature

genus

the category of the feature

Examples

data(cheese)
head(annotations)
#> # A tibble: 6 × 9
#>   ID    GO_ID KEGG_ID domain   phylum         class           order family genus
#>   <chr> <chr> <chr>   <chr>    <chr>          <chr>           <chr> <chr>  <chr>
#> 1 9     -     -       Bacteria Pseudomonadota Alphaproteobac… Caul… Caulo… Caul…
#> 2 10    -     -       Bacteria Pseudomonadota Alphaproteobac… Caul… Caulo… Caul…
#> 3 11    -     -       Bacteria Pseudomonadota Alphaproteobac… Hyph… Barto… Bart…
#> 4 12    -     -       Bacteria Pseudomonadota Alphaproteobac… Hyph… Bruce… Bruc…
#> 5 13    -     -       Bacteria Pseudomonadota Alphaproteobac… Hyph… Nitro… Brad…
#> 6 14    -     -       Bacteria Pseudomonadota Alphaproteobac… Hyph… Phyll… Meso…
head(cheese)
#> # A tibble: 6 × 18
#>   ID      A_1   A_2   A_3   A_4   A_5   C_1   C_3   C_4   C_5 GO_ID KEGG_ID
#>   <chr> <int> <int> <int> <int> <int> <int> <int> <int> <int> <chr> <chr>  
#> 1 1        38    23    27     5    11    13     1     3     0 NA    NA     
#> 2 2        23     6     4     0     9    21     1     0     2 NA    NA     
#> 3 3        24     5    37    10    19     3     0     1     4 NA    NA     
#> 4 4         3     2     4     2    16    56     1     7    17 NA    NA     
#> 5 5        58    14    45    13    32    82     3     8     3 NA    NA     
#> 6 6        12     9    14     4    13     2     0     1     4 NA    NA     
#> # ℹ 6 more variables: domain <chr>, phylum <chr>, class <chr>, order <chr>,
#> #   family <chr>, genus <chr>