Cheese data

In the context of cheese production, the regular application of a brine solution is a maturation technique that promotes uniformity in the microbial populations on the cheese's surface and facilitates interactions among these microorganisms. This investigation involved the analysis of a longitudinal dataset encompassing three washed-rind cheese communities sampled during the ripening process.

Format

two datasets: cheese and annotations

Source

Reference: doi:10.1128/msystems.00701-22

Details

The study uncovered a remarkably consistent microbial progression within each cheese. In the bacterial community, Firmicutes dominate at the outset, with Proteobacteria swiftly assuming dominance by the end of the ripening period. Additionally, both Cheese A and Cheese C consistently demonstrate the establishment of Actinobacteria and Bacteroidetes, each in its own distinct manner. To corroborate these findings using the MolPad dashboard, we conducted an analysis of two cheeses (A and C) from all three production batches, spanning weeks 2 to 13.

cheese data

A data frame with 29423 rows and 11 variables:

ID

sample ID

A1~A5

Time series data measured for cheese 1.

B1~B5

Time series data measured for cheese 2.

annotation data

A data frame with 29423 rows and 10 variables:

ID

sample ID

GO_ID

GO IDs, represents a link between a gene product type and a molecular function

KEGG_ID

KEGG IDs, linking genomic information with higher order functional information.

Other_ID

other IDs in online databases

kingdom

the category of the feature

phylum

the category of the feature

class

the category of the feature

order

the category of the feature

family

the category of the feature

genus

the category of the feature

Examples

data(cheese)
head(annotations)
#>       ID GO_ID KEGG_ID               Other_ID   kingdom         phylum
#> 1      1                              DUF1329  Bacteria Proteobacteria
#> 2   1000                                 FliJ  Bacteria Proteobacteria
#> 3  10001        K07304                   PMSR  Bacteria     Firmicutes
#> 4 100040                             ABC_tran  Bacteria Proteobacteria
#> 5 100058                ABC_membrane,ABC_tran Eukaryota   Opisthokonta
#> 6 100067        K05685 ABC_tran,FtsX,MacB_PCD  Bacteria Proteobacteria
#>                 class                  order family genus
#> 1 Gammaproteobacteria Pseudoalteromonadaceae   <NA>  <NA>
#> 2 Gammaproteobacteria                   <NA>   <NA>  <NA>
#> 3             Bacilli            Lactococcus   <NA>  <NA>
#> 4  Betaproteobacteria                   <NA>   <NA>  <NA>
#> 5               Fungi             Ascomycota   <NA>  <NA>
#> 6 Gammaproteobacteria          Moraxellaceae   <NA>  <NA>
head(cheese)
#>       ID A1 A2 A3 A4 A5 B1 B2 B3 B4 B5
#> 1      1  6  1  1  0  0  4  4  4  4  4
#> 2   1000  1  0  0  0  0  3  2  1  1  1
#> 3  10001  2  0  0  0  1  0  1  0  0  0
#> 4 100040  0  0  0  0  1  0  0  0  0  1
#> 5 100058  0  0  0  0  1  1  0  0  0  0
#> 6 100067  0  0  0  0  1  2  0  0  2  0